source/frontend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
stomp.py==8.1.0
httpx==0.26.0
jinja2==3.1.3
python-multipart==0.0.6
websockets
requests
